+++ src/app/layout/transport/transport.component.ts
@@ -3,10 +3,10 @@ import { Router } from '@angular/router';
 import { TransportService } from '../../core/services/transport.service';
 import { ToastrService } from 'ngx-toastr';
 import { NgxSpinnerService } from 'ngx-spinner';
-
 import { HelpService } from '../../core/services/help.service';
 import * as Globals from '../../core/globals';
-
+import { MatDialog, MatDialogRef } from '@angular/material';
+import { ConfirmDialogComponent } from '../../core/component/confirm-dialog/confirm-dialog.component';
 @Component({
   selector: 'app-transport',
   templateUrl: './transport.component.html',
@@ -24,12 +24,14 @@ export class TransportComponent implements OnInit {
   upper_count: number;
   paginationMaxSize: number;
   itemPerPage: number;
+  dialogRef: MatDialogRef<ConfirmDialogComponent>;
   constructor(
     private router: Router,
     private toastr: ToastrService,
     private transportService: TransportService,
     private spinner: NgxSpinnerService,
-    private helpService: HelpService
+    private helpService: HelpService,
+    public dialog: MatDialog
   ) { }
 
   ngOnInit() {
@@ -48,8 +50,8 @@ export class TransportComponent implements OnInit {
     })
   }
 
-  btnClickNav= function (toNav) {
-    this.router.navigateByUrl('/'+toNav);
+  btnClickNav = function (toNav) {
+    this.router.navigateByUrl('/' + toNav);
   };
 
   dataSearch() {
@@ -57,29 +59,29 @@ export class TransportComponent implements OnInit {
     this.defaultPagination = 1;
     this.getTransportList();
   }
-  
+
   getTransportList() {
     let params: URLSearchParams = new URLSearchParams();
     params.set('page', this.defaultPagination.toString());
     params.set('search', this.search_key.toString());
     this.transportService.getTransporterList(params).subscribe(
       (data: any[]) => {
-       
+
         this.totalTransportList = data['count'];
         this.transportList = data['results'];
         this.itemNo = (this.defaultPagination - 1) * this.itemPerPage;
         this.lower_count = this.itemNo + 1;
-        if(this.totalTransportList > this.itemPerPage*this.defaultPagination){
-          this.upper_count = this.itemPerPage*this.defaultPagination
+        if (this.totalTransportList > this.itemPerPage * this.defaultPagination) {
+          this.upper_count = this.itemPerPage * this.defaultPagination
         }
-        else{
+        else {
           this.upper_count = this.totalTransportList
         }
         this.spinner.hide();
       }
     );
   };
-  
+
   activeState(id) {
     this.spinner.show();
     let transporter;
@@ -130,27 +132,38 @@ export class TransportComponent implements OnInit {
   };
 
   deletetransport(id) {
-    this.spinner.show();
-    let transporter;
+    this.dialogRef = this.dialog.open(ConfirmDialogComponent, {
+      disableClose: false
+    });
+    this.dialogRef.componentInstance.confirmMessage = "Are you sure you want to delete?"
 
-    transporter = {
-      id: id
-    };
+    this.dialogRef.afterClosed().subscribe(result => {
+      if (result) {
+        this.spinner.show();
+        let transporter;
 
-    this.transportService.deleteTransporter(transporter).subscribe(
-      response => {
-        this.toastr.success('Transporter deleted successfully', '', {
-          timeOut: 3000,
-        });
-        this.getTransportList();
-      },
-      error => {
-        console.log('error', error)
-        // this.toastr.error('everything is broken', '', {
-        //   timeOut: 3000,
-        // });
+        transporter = {
+          id: id
+        };
+
+        this.transportService.deleteTransporter(transporter).subscribe(
+          response => {
+            this.toastr.success('Transporter deleted successfully', '', {
+              timeOut: 3000,
+            });
+            this.getTransportList();
+          },
+          error => {
+            console.log('error', error)
+            // this.toastr.error('everything is broken', '', {
+            //   timeOut: 3000,
+            // });
+          }
+        );
       }
-    );
+      this.dialogRef = null;
+    });
+
   };
 
   pagination() {
